requirements.txt
# PyTorch
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
torchvision==0.13.1+cu113

# Lightning + Hydra
lightning==2.0.1
pytorch_lightning==2.0.9
torchmetrics==1.0.3
hydra-core==1.3
hydra-zen==0.12.1
hydra_colorlog==1.2.0
black
setuptools>=68.0

# Common utilities
tensorboardX
opencv-python
numpy==1.23.5
ipdb
ffmpeg-python
jupyter
matplotlib
scikit-image
termcolor
imageio==2.14.1
imageio-ffmpeg==0.4.7
einops==0.6.1

# tensorbordX save video needs moviepy
moviepy

# Diffusion
diffusers[torch]==0.19.3
transformers==4.31.0
timm==0.9.12  # For HMR2.0a feature extraction

# 3D-Vision
pytorch3d @ https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu113_pyt1121/pytorch3d-0.7.2-cp38-cp38-linux_x86_64.whl
trimesh
chumpy
smplx
open3d==0.17.0
pyrender

# ray
termcolor
ray==2.6.1
pydantic==1.10.11  # TODO: FIX after ray 2.7

# Wis3d
wis3d

# New
decord

spacy==3.7.6